move generic issue-metadata into DoFormBase
